posts/utils/permissions.py
from django.contrib.auth import get_user_model

from rest_framework.permissions import BasePermission, SAFE_METHODS

class IsOwnerOrReadOnly(BasePermission):
    message = "해당 게시글의 권한이 없습니다."
    def has_object_permission(self, request, view, obj):
        if request.method in SAFE_METHODS:
            return True
        else:
            return obj.user == request.user
        
    
    